resolve database seeding issues and add docker scripts
