+++ db/seeds/b_seed.js
@@ -1,27 +1,9 @@
-// ADD ADDITIONAL IMPORTS FOR SAMPLE_SEEDS HERE
-
-
-
-var users = require('../sample_seeds/users');
-var locations = require('../sample_seeds/locations');
 var locationsusers = require('../sample_seeds/locationsusers');
 var connections = require('../sample_seeds/connections');
 var channels = require('../sample_seeds/channels');
 
 exports.seed = (knex, Promise) => {
   var allPromises = [];
-
-  // ADD ADDITIONAL FOREACH LOOPS HERE
-
-
-
-  users.forEach((user) => {
-    allPromises.push(createUser(knex, user));
-  });
-
-  locations.forEach(function(location){
-    allPromises.push(createLocation(knex, location));
-  });
 
   locationsusers.forEach(function(locationuser){
     allPromises.push(createLocationUser(knex, locationuser));
@@ -35,43 +17,7 @@ exports.seed = (knex, Promise) => {
     allPromises.push(createChannel(knex, channel));
   });
 
-
   return Promise.all(allPromises);
-};
-
-
-
-// ADD ADDITIONAL FUNCTIONS HERE
-
-
-
-function createUser(knex, user) {
-  return knex.table('users')
-    .returning('id')
-    .insert({
-      first: user.first,
-      last: user.last,
-      email: user.email
-    })
-    .catch((error) => {
-      console.log('ERROR:', error);
-    });
-};
-
-function createLocation(knex, location) {
-  return knex.table('locations')
-    .returning('id')
-    .insert({
-      category: location.category,
-      latitude: location.latitude,
-      longitude: location.longitude,
-      name: location.name,
-      city: location.city,
-      state: location.state
-    })
-    .catch((error) => {
-      console.log('ERROR:', error);
-    });
 };
 
 function createLocationUser(knex, locationuser) {
